[data] refactor datapipe
